Added tests for vault-multistaking integration for auto-stake feature (#33)
