+++ contracts/vault/Cargo.toml
@@ -42,6 +42,8 @@ weighted-pool = { path = "../weighted_pool"}
 stableswap-pool = { path = "../stable_pool"}
 xyk-pool = { path = "../xyk_pool"}
 lp-token = { path = "../lp_token"}
+dexter-generator = { path = "../dexter_generator/generator"}
+dexter-multi-staking = { path = "../dexter_generator/multi_staking"}
 
 cw-multi-test = "0.13.2"
 cw20 = { version = "0.15" }
