+++ contracts/vault/tests/owner.rs
@@ -1,0 +1,113 @@
+pub mod utils;
+
+use cosmwasm_std::Addr;
+use cw_multi_test::Executor;
+use dexter::vault::{ConfigResponse, ExecuteMsg, QueryMsg};
+use crate::utils::{instantiate_contract, mock_app};
+
+#[test]
+fn test_update_owner() {
+    let owner = String::from("owner");
+    let mut app = mock_app(Addr::unchecked(owner.clone()), vec![]);
+    let vault_instance = instantiate_contract(&mut app, &Addr::unchecked(owner.clone()));
+
+    let new_owner = String::from("new_owner");
+
+    // New owner
+    let msg = ExecuteMsg::ProposeNewOwner {
+        owner: new_owner.clone(),
+        expires_in: 100, // seconds
+    };
+
+    // Unauthed check
+    let err = app
+        .execute_contract(
+            Addr::unchecked("not_owner"),
+            vault_instance.clone(),
+            &msg,
+            &[],
+        )
+        .unwrap_err();
+    assert_eq!(err.root_cause().to_string(), "Generic error: Unauthorized");
+
+    // Claim before proposal
+    let err = app
+        .execute_contract(
+            Addr::unchecked(new_owner.clone()),
+            vault_instance.clone(),
+            &ExecuteMsg::ClaimOwnership {},
+            &[],
+        )
+        .unwrap_err();
+    assert_eq!(
+        err.root_cause().to_string(),
+        "Generic error: Ownership proposal not found"
+    );
+
+    // Propose new owner
+    app.execute_contract(Addr::unchecked("owner"), vault_instance.clone(), &msg, &[])
+        .unwrap();
+
+    // Claim from invalid addr
+    let err = app
+        .execute_contract(
+            Addr::unchecked("invalid_addr"),
+            vault_instance.clone(),
+            &ExecuteMsg::ClaimOwnership {},
+            &[],
+        )
+        .unwrap_err();
+    assert_eq!(err.root_cause().to_string(), "Generic error: Unauthorized");
+
+    // Drop ownership proposal
+    let err = app
+        .execute_contract(
+            Addr::unchecked(new_owner.clone()),
+            vault_instance.clone(),
+            &ExecuteMsg::DropOwnershipProposal {},
+            &[],
+        )
+        .unwrap_err();
+    // new_owner is not an owner yet
+    assert_eq!(err.root_cause().to_string(), "Generic error: Unauthorized");
+
+    app.execute_contract(
+        Addr::unchecked(owner.clone()),
+        vault_instance.clone(),
+        &ExecuteMsg::DropOwnershipProposal {},
+        &[],
+    )
+    .unwrap();
+
+    // Try to claim ownership
+    let err = app
+        .execute_contract(
+            Addr::unchecked(new_owner.clone()),
+            vault_instance.clone(),
+            &ExecuteMsg::ClaimOwnership {},
+            &[],
+        )
+        .unwrap_err();
+    assert_eq!(
+        err.root_cause().to_string(),
+        "Generic error: Ownership proposal not found"
+    );
+
+    // Propose new owner again
+    app.execute_contract(Addr::unchecked("owner"), vault_instance.clone(), &msg, &[])
+        .unwrap();
+    // Claim ownership
+    app.execute_contract(
+        Addr::unchecked(new_owner.clone()),
+        vault_instance.clone(),
+        &ExecuteMsg::ClaimOwnership {},
+        &[],
+    )
+    .unwrap();
+
+    // Let's query the contract state
+    let msg = QueryMsg::Config {};
+    let res: ConfigResponse = app.wrap().query_wasm_smart(&vault_instance, &msg).unwrap();
+
+    assert_eq!(res.owner, new_owner)
+}
