+++ contracts/vault/tests/pool_creation.rs
@@ -1,0 +1,531 @@
+pub mod utils;
+
+use cosmwasm_std::{attr, coin, Addr, Coin, Uint128};
+use cw20::MinterResponse;
+use cw_multi_test::Executor;
+use dexter::asset::{Asset, AssetInfo};
+use dexter::lp_token::InstantiateMsg as TokenInstantiateMsg;
+
+use dexter::vault::{AllowPoolInstantiation, ExecuteMsg, PoolInfo, PoolType, QueryMsg};
+
+use crate::utils::{dummy_pool_creation_msg, instantiate_contract, mock_app, store_token_code};
+
+#[test]
+fn test_create_pool_instance() {
+    let owner = String::from("owner");
+    let mut app = mock_app(
+        Addr::unchecked(owner.clone()),
+        vec![Coin {
+            denom: "uxprt".to_string(),
+            amount: Uint128::from(1_000_000_000u64),
+        }],
+    );
+
+    let owner_addr = Addr::unchecked(owner.clone());
+    let user_addr = Addr::unchecked("user".to_string());
+
+    // Send some funds from owner to user
+    app.send_tokens(
+        owner_addr.clone(),
+        user_addr.clone(),
+        &[coin(200_000_000u128, "uxprt")],
+    )
+    .unwrap();
+
+    let token_code_id = store_token_code(&mut app);
+    let vault_instance = instantiate_contract(&mut app, &Addr::unchecked(owner.clone()));
+
+    // Create Token X
+    let init_msg = TokenInstantiateMsg {
+        name: "x_token".to_string(),
+        symbol: "X-Tok".to_string(),
+        decimals: 18,
+        initial_balances: vec![],
+        mint: Some(MinterResponse {
+            minter: owner.to_string(),
+            cap: None,
+        }),
+        marketing: None,
+    };
+    let token_instance0 = app
+        .instantiate_contract(
+            token_code_id,
+            Addr::unchecked(owner.clone()),
+            &init_msg,
+            &[],
+            "x_token",
+            None,
+        )
+        .unwrap();
+
+    // Create Token Y
+    let init_msg = TokenInstantiateMsg {
+        name: "y_token".to_string(),
+        symbol: "Y-Tok".to_string(),
+        decimals: 18,
+        initial_balances: vec![],
+        mint: Some(MinterResponse {
+            minter: owner.to_string(),
+            cap: None,
+        }),
+        marketing: None,
+    };
+    let token_instance1 = app
+        .instantiate_contract(
+            token_code_id,
+            Addr::unchecked(owner.clone()),
+            &init_msg,
+            &[],
+            "y_token",
+            None,
+        )
+        .unwrap();
+
+    let asset_infos = vec![
+        AssetInfo::Token {
+            contract_addr: token_instance0.clone(),
+        },
+        AssetInfo::Token {
+            contract_addr: token_instance1.clone(),
+        },
+    ];
+
+    let msg = ExecuteMsg::CreatePoolInstance {
+        pool_type: PoolType::Xyk {},
+        asset_infos: asset_infos.to_vec(),
+        init_params: None,
+        fee_info: None,
+    };
+
+    let res = app
+        .execute_contract(
+            Addr::unchecked(owner.clone()),
+            vault_instance.clone(),
+            &msg,
+            &[],
+        )
+        .unwrap();
+
+    assert_eq!(res.events[1].attributes[2], attr("pool_type", "xyk"));
+
+    let pool_res: PoolInfo = app
+        .wrap()
+        .query_wasm_smart(
+            vault_instance.clone(),
+            &QueryMsg::GetPoolById {
+                pool_id: Uint128::from(1u128),
+            },
+        )
+        .unwrap();
+
+    let assets = vec![
+        Asset {
+            info: AssetInfo::Token {
+                contract_addr: token_instance0.clone(),
+            },
+            amount: Uint128::zero(),
+        },
+        Asset {
+            info: AssetInfo::Token {
+                contract_addr: token_instance1.clone(),
+            },
+            amount: Uint128::zero(),
+        },
+    ];
+
+    assert_eq!(Uint128::from(1u128), pool_res.pool_id);
+    assert_eq!(
+        Addr::unchecked("contract3".to_string()),
+        pool_res.lp_token_addr
+    );
+    assert_eq!(
+        Addr::unchecked("contract4".to_string()),
+        pool_res.pool_addr
+    );
+    assert_eq!(assets, pool_res.assets);
+    assert_eq!(PoolType::Xyk {}, pool_res.pool_type);
+}
+
+#[test]
+fn test_pool_creation_whitelist() {
+    let owner = String::from("owner");
+    let mut app = mock_app(
+        Addr::unchecked(owner.clone()),
+        vec![Coin {
+            denom: "uxprt".to_string(),
+            amount: Uint128::from(1_000_000_000u64),
+        }],
+    );
+
+    let owner_addr = Addr::unchecked(owner.clone());
+    let user_addr = Addr::unchecked("user".to_string());
+
+    // Send some funds from owner to user
+    app.send_tokens(
+        owner_addr.clone(),
+        user_addr.clone(),
+        &[coin(300_000_000u128, "uxprt")],
+    )
+    .unwrap();
+
+    let token_code_id = store_token_code(&mut app);
+    let vault_instance = instantiate_contract(&mut app, &Addr::unchecked(owner.clone()));
+
+    // Create Token X
+    let init_msg = TokenInstantiateMsg {
+        name: "x_token".to_string(),
+        symbol: "X-Tok".to_string(),
+        decimals: 18,
+        initial_balances: vec![],
+        mint: Some(MinterResponse {
+            minter: owner.to_string(),
+            cap: None,
+        }),
+        marketing: None,
+    };
+    let token_instance0 = app
+        .instantiate_contract(
+            token_code_id,
+            Addr::unchecked(owner.clone()),
+            &init_msg,
+            &[],
+            "x_token",
+            None,
+        )
+        .unwrap();
+
+    // Create Token Y
+    let init_msg = TokenInstantiateMsg {
+        name: "y_token".to_string(),
+        symbol: "Y-Tok".to_string(),
+        decimals: 18,
+        initial_balances: vec![],
+        mint: Some(MinterResponse {
+            minter: owner.to_string(),
+            cap: None,
+        }),
+        marketing: None,
+    };
+    let token_instance1 = app
+        .instantiate_contract(
+            token_code_id,
+            Addr::unchecked(owner.clone()),
+            &init_msg,
+            &[],
+            "y_token",
+            None,
+        )
+        .unwrap();
+
+    let asset_infos = vec![
+        AssetInfo::Token {
+            contract_addr: token_instance0.clone(),
+        },
+        AssetInfo::Token {
+            contract_addr: token_instance1.clone(),
+        },
+    ];
+
+    // Set a pool creation fee
+    let msg = ExecuteMsg::UpdateConfig {
+        lp_token_code_id: None,
+        fee_collector: None,
+        pool_creation_fee: Some(Asset {
+            info: AssetInfo::NativeToken {
+                denom: "uxprt".to_string(),
+            },
+            amount: Uint128::from(100_000_000u128),
+        }),
+        auto_stake_impl: None,
+        multistaking_address: None,
+        generator_address: None,
+    };
+
+    app.execute_contract(
+        Addr::unchecked(owner.clone()),
+        vault_instance.clone(),
+        &msg,
+        &[],
+    )
+    .unwrap();
+
+    // Pool creation allowed to anyone scenarion
+    let msg = dummy_pool_creation_msg(&asset_infos);
+    let res = app.execute_contract(
+        user_addr.clone(),
+        vault_instance.clone(),
+        &msg,
+        &[coin(100_000_000u128, "uxprt")],
+    );
+
+    assert!(res.is_ok());
+
+    // disable pool creation for everyone
+    let msg = ExecuteMsg::UpdatePoolTypeConfig {
+        pool_type: PoolType::Xyk {},
+        allow_instantiation: Some(AllowPoolInstantiation::Nobody),
+        is_generator_disabled: None,
+        new_fee_info: None,
+    };
+
+    app.execute_contract(
+        Addr::unchecked(owner.clone()),
+        vault_instance.clone(),
+        &msg,
+        &[],
+    )
+    .unwrap();
+
+    // try creating another pool
+    let msg = dummy_pool_creation_msg(&asset_infos);
+    let res = app.execute_contract(
+        Addr::unchecked(owner.clone()),
+        vault_instance.clone(),
+        &msg,
+        &[coin(100_000_000u128, "uxprt")],
+    );
+
+    assert!(res.is_err());
+    assert_eq!(
+        res.unwrap_err().root_cause().to_string(),
+        "Creation of this pool type is disabled"
+    );
+
+    // enable pool creation for only whitelisted addresses
+    let msg = ExecuteMsg::UpdatePoolTypeConfig {
+        pool_type: PoolType::Xyk {},
+        allow_instantiation: Some(AllowPoolInstantiation::OnlyWhitelistedAddresses),
+        is_generator_disabled: None,
+        new_fee_info: None,
+    };
+
+    app.execute_contract(
+        Addr::unchecked(owner.clone()),
+        vault_instance.clone(),
+        &msg,
+        &[],
+    )
+    .unwrap();
+
+    // Pool instantiation from admin
+    let msg = dummy_pool_creation_msg(&asset_infos);
+    let res = app.execute_contract(
+        Addr::unchecked(owner.clone()),
+        vault_instance.clone(),
+        &msg,
+        &[coin(100_000_000u128, "uxprt")],
+    );
+
+    // Pool instantiation from admin should work regardless of empty whitelist
+    assert!(res.is_ok());
+
+    // Pool instantiation from non-admin non-whitelisted address
+    let msg = dummy_pool_creation_msg(&asset_infos);
+    let res = app.execute_contract(
+        user_addr.clone(),
+        vault_instance.clone(),
+        &msg,
+        &[coin(100_000_000u128, "uxprt")],
+    );
+
+    assert!(res.is_err());
+    assert_eq!(res.unwrap_err().root_cause().to_string(), "Unauthorized");
+
+    // Add user to whitelist
+    let msg = ExecuteMsg::AddAddressToWhitelist {
+        address: user_addr.to_string(),
+    };
+    app.execute_contract(
+        Addr::unchecked(owner.clone()),
+        vault_instance.clone(),
+        &msg,
+        &[],
+    )
+    .unwrap();
+
+    // Pool instantiation from non-admin whitelisted address
+    let msg = dummy_pool_creation_msg(&asset_infos);
+    let res = app.execute_contract(
+        user_addr.clone(),
+        vault_instance.clone(),
+        &msg,
+        &[coin(100_000_000u128, "uxprt")],
+    );
+
+    assert!(res.is_ok());
+
+    // Remove user from whitelist and test again
+    let msg = ExecuteMsg::RemoveAddressFromWhitelist {
+        address: user_addr.to_string(),
+    };
+    app.execute_contract(
+        Addr::unchecked(owner.clone()),
+        vault_instance.clone(),
+        &msg,
+        &[],
+    )
+    .unwrap();
+
+    // Pool instantiation from non-admin non-whitelisted address
+    let msg = dummy_pool_creation_msg(&asset_infos);
+    let res = app.execute_contract(
+        user_addr.clone(),
+        vault_instance.clone(),
+        &msg,
+        &[coin(100_000_000u128, "uxprt")],
+    );
+
+    assert!(res.is_err());
+    assert_eq!(res.unwrap_err().root_cause().to_string(), "Unauthorized");
+}
+
+#[test]
+fn test_pool_creation_fee() {
+    let owner = String::from("owner");
+    let mut app = mock_app(
+        Addr::unchecked(owner.clone()),
+        vec![Coin {
+            denom: "uxprt".to_string(),
+            amount: Uint128::from(1_000_000_000u64),
+        }],
+    );
+
+    let owner_addr = Addr::unchecked(owner.clone());
+    let user_addr = Addr::unchecked("user".to_string());
+
+    // Send some funds from owner to user
+    app.send_tokens(
+        owner_addr.clone(),
+        user_addr.clone(),
+        &[coin(200_000_000u128, "uxprt")],
+    )
+    .unwrap();
+
+    let token_code_id = store_token_code(&mut app);
+    let vault_instance = instantiate_contract(&mut app, &Addr::unchecked(owner.clone()));
+
+    // Create Token X
+    let init_msg = TokenInstantiateMsg {
+        name: "x_token".to_string(),
+        symbol: "X-Tok".to_string(),
+        decimals: 18,
+        initial_balances: vec![],
+        mint: Some(MinterResponse {
+            minter: owner.to_string(),
+            cap: None,
+        }),
+        marketing: None,
+    };
+    let token_instance0 = app
+        .instantiate_contract(
+            token_code_id,
+            Addr::unchecked(owner.clone()),
+            &init_msg,
+            &[],
+            "x_token",
+            None,
+        )
+        .unwrap();
+
+    // Create Token Y
+    let init_msg = TokenInstantiateMsg {
+        name: "y_token".to_string(),
+        symbol: "Y-Tok".to_string(),
+        decimals: 18,
+        initial_balances: vec![],
+        mint: Some(MinterResponse {
+            minter: owner.to_string(),
+            cap: None,
+        }),
+        marketing: None,
+    };
+    let token_instance1 = app
+        .instantiate_contract(
+            token_code_id,
+            Addr::unchecked(owner.clone()),
+            &init_msg,
+            &[],
+            "y_token",
+            None,
+        )
+        .unwrap();
+
+    let asset_infos = vec![
+        AssetInfo::Token {
+            contract_addr: token_instance0.clone(),
+        },
+        AssetInfo::Token {
+            contract_addr: token_instance1.clone(),
+        },
+    ];
+
+    // No pool creation fee
+    let msg = ExecuteMsg::CreatePoolInstance {
+        pool_type: PoolType::Xyk {},
+        asset_infos: asset_infos.to_vec(),
+        init_params: None,
+        fee_info: None,
+    };
+
+    let res = app
+        .execute_contract(
+            Addr::unchecked(owner.clone()),
+            vault_instance.clone(),
+            &msg,
+            &[],
+        )
+        .unwrap();
+
+    assert_eq!(res.events[1].attributes[2], attr("pool_type", "xyk"));
+
+    // Add fee for pool creation
+    let msg = ExecuteMsg::UpdateConfig {
+        lp_token_code_id: None,
+        fee_collector: None,
+        pool_creation_fee: Some(Asset {
+            info: AssetInfo::NativeToken {
+                denom: "uxprt".to_string(),
+            },
+            amount: Uint128::from(100_000_000u128),
+        }),
+        auto_stake_impl: None,
+        generator_address: None,
+        multistaking_address: None,
+    };
+
+    app.execute_contract(
+        Addr::unchecked(owner.clone()),
+        vault_instance.clone(),
+        &msg,
+        &[],
+    )
+    .unwrap();
+
+    // try creating another pool without passing fee
+    let msg = dummy_pool_creation_msg(&asset_infos);
+    let res = app.execute_contract(
+        Addr::unchecked(owner.clone()),
+        vault_instance.clone(),
+        &msg,
+        &[],
+    );
+
+    assert!(res.is_err());
+    assert_eq!(
+        res.unwrap_err().root_cause().to_string(),
+        "Insufficient number of uxprt tokens sent. Tokens sent = 0. Tokens needed = 100000000"
+    );
+
+    // try creating another pool with passing fee
+    let msg = dummy_pool_creation_msg(&asset_infos);
+    let res = app
+        .execute_contract(
+            Addr::unchecked(owner.clone()),
+            vault_instance.clone(),
+            &msg,
+            &[coin(100_000_000u128, "uxprt")],
+        )
+        .unwrap();
+
+    assert_eq!(res.events[1].attributes[2], attr("pool_type", "xyk"));
+}
